config: share lua loader
